refactor(orgstore): extract org/tenant subsystem into importable package
Extract Organizations/Teams/Memberships/Roles/Invitations/AccessGrants + RBAC
authorizer from *PostgresStore and *HelixAPIServer into api/pkg/orgstore, which
operates on a plain *gorm.DB so it can be vendored by downstream services
(HelixOS). PostgresStore embeds *orgstore.Store (methods promoted); query structs
are aliased in the store package so existing call sites are unchanged; the server
authz helpers delegate to orgstore.Authorizer, removing the duplicated
getAuthzConfigs/evaluate. No behaviour change.

Spec-Ref: helix-specs@d2d0edf:002222_make-helixos-modular-so
